+++ packages/dumb/src/components/tooltips/warm-up-tooltip/warm-up-tooltip.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Diabeloop
+ * Copyright (c) 2023-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -29,7 +29,6 @@ import { Tooltip } from '../../../index'
 import React, { FC } from 'react'
 import { TimePrefs } from 'medical-domain'
 import {
-  COMMON_TOOLTIP_SIDE,
   COMMON_TOOLTIP_TAIL_HEIGHT,
   COMMON_TOOLTIP_TAIL_WIDTH,
   DEFAULT_TOOLTIP_BORDER_WIDTH,
@@ -62,7 +61,7 @@ export const WarmUpTooltip: FC<WarmupTooltipProps> = (props) => {
         position={position}
         borderColor={colors.deviceEvent}
         dateTitle={getDateTitleForBaseDatum(warmup, timePrefs)}
-        side={side || COMMON_TOOLTIP_SIDE}
+        side={side}
         tailWidth={COMMON_TOOLTIP_TAIL_WIDTH}
         tailHeight={COMMON_TOOLTIP_TAIL_HEIGHT}
         tail={DEFAULT_TOOLTIP_TAIL}