+++ packages/dumb/src/index.ts
@@ -76,6 +76,8 @@ import { Device } from './models/device.model'
 import { buildDevice } from './utils/device/device.utils'
 import { SafetyBasalItem } from './models/safety-basal-item.model'
 import { isSafetyBasalAvailable, getSafetyBasalItems } from './utils/safety-basal-profile/safety-basal-profile.util'
+import { NightModeTooltip } from './components/tooltips/night-mode-tooltip/night-mode-tooltip'
+import { SimpleValue } from './components/stats/common/simple-value'
 
 export {
   formatDate,
@@ -105,6 +107,7 @@ export {
   formatBgValue,
   LayoutColumnType,
   LoopModeStat,
+  NightModeTooltip,
   ParameterTooltip,
   PhysicalTooltip,
   PrintView,
@@ -124,5 +127,6 @@ export {
   WarmUpTooltip,
   SafetyBasalItem,
   isSafetyBasalAvailable,
-  getSafetyBasalItems
+  getSafetyBasalItems,
+  SimpleValue
 }