+++ packages/yourloops/lib/team/index.ts
@@ -26,12 +26,10 @@
  */
 
 import { type TeamMember } from './models/team-member.model'
-import { REGEX_TEAM_CODE, REGEX_TEAM_CODE_DISPLAY, type Team, TEAM_CODE_LENGTH } from './models/team.model'
+import { REGEX_TEAM_CODE, REGEX_TEAM_CODE_DISPLAY, type Team } from './models/team.model'
 import { type TeamContext } from './models/team-context.model'
 
 import { TeamContextProvider, useTeam } from './team.hook'
-
-import { getDisplayTeamCode } from './team.util'
 
 export {
   type Team,
@@ -39,8 +37,6 @@ export {
   type TeamContext,
   TeamContextProvider,
   useTeam,
-  getDisplayTeamCode,
-  TEAM_CODE_LENGTH,
   REGEX_TEAM_CODE,
   REGEX_TEAM_CODE_DISPLAY
 }