+++ test/locales/translations.test.js
@@ -71,7 +71,11 @@ const ignoredTransKeysForBlip = [
   'Parameter',
   'Parameters History',
   'Reservoir changes',
-  'Safety basal profile'
+  'Safety basal profile',
+  'millisecond',
+  'milliseconds',
+  'days',
+  'second'
 ]
 const ignoredTransKeyInBlipFiles = [
   '${physicalActivity.reportedIntensity}-pa',
@@ -82,7 +86,9 @@ const ignoredTransKeyInBlipFiles = [
   '${titleType} Above Range',
   '${titleType} Below Range',
   '${titleType} In Range',
-  'Micro Bolus'
+  'Micro Bolus',
+  'params|${nameUppercase}',
+  'params|${PhysicalActivityName.AerobicDefault}'
 ]
 const ignoredTransKeyForYourLoops = [
   // Countries (from locales/languages.json)
@@ -154,6 +160,7 @@ const ignoredTransKeyForYourLoops = [
   // Others
   // TODO
   'accompanying-documents',
+  'account-flagged-for-deletion',
   'alarm-alert-loop-mode-activated-description',
   'alarm-alert-loop-mode-deactivated-description',
   'alarm-alert-with-code',
@@ -231,6 +238,7 @@ const ignoredTransKeyForYourLoops = [
   'login-page-desktop-info-3',
   'login-page-desktop-title',
   'login-page-mobile-title',
+  'night-mode',
   'end-date',
   'remaining-time',
   'time-in-range-cgm-daily-average',
@@ -329,7 +337,13 @@ const ignoredTransKeyInYourLoopsFiles = [
   'Serial Number',
   'Setting',
   'Settings on day',
-  'Time In Range'
+  'Time In Range',
+  'Infusion site changes',
+  'days',
+  'hours',
+  'minutes',
+  'seconds',
+  'milliseconds'
 ]
 
 /**